- m_stats.c:stats_events(): use proper conversion specifier
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/trunk@7458 82007160-df01-0410-b94d-b575c5fd34c7

+++ modules/m_stats.c
@@ -603,9 +603,8 @@ stats_events(struct Client *source_p, int parc, char *parv[])
     const struct event *ev = node->data;
 
     sendto_one_numeric(source_p, &me, RPL_STATSDEBUG | SND_EXPLICIT,
-                       "E :%-30s %-4d seconds",
-                       ev->name,
-                       (int)(ev->next - CurrentTime));
+                       "E :%-30s %-4ji seconds",
+                       ev->name, ev->next - CurrentTime);
   }
 }
